feat: move AI Fill from quiz header to exam detail page
- Remove startFill, fill state, and Sparkles button from QuizHeader
  (it was appearing on every quiz/review/answers page, which is unintuitive)
- Add AI Fill button to ExamDetailClient, below the Translate button,
  with inline progress display (filling N/total, filled X skipped Y)
- Also fix settings-context: guard aiFillPrompt and aiFillPromptVersions
  with defaults during merge, matching the pattern for other prompts
